Move order operation messages to separate component
